refactored accounts: created new app for authentication
