+++ reservation/views.py
@@ -1,13 +1,12 @@
 from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.contrib.auth.views import LoginView, LogoutView
 from django.contrib.messages.views import SuccessMessageMixin
 from django.db.models import QuerySet
-from django.urls import reverse_lazy, reverse
-from django.views.generic import ListView, CreateView, DetailView, UpdateView
+from django.urls import reverse
+from django.views.generic import ListView, CreateView, DetailView
 
-from reservation.forms import VisitorCreationForm, VisitorUpdateForm, ReservationForm
-from reservation.models import Cafe, City, Visitor, Reservation, Table
+from reservation.forms import ReservationForm
+from reservation.models import Cafe, City, Reservation, Table
 
 
 class CafeListView(LoginRequiredMixin, ListView):
@@ -166,37 +165,3 @@ class ReservationListView(LoginRequiredMixin, ListView):
 
     def get_queryset(self):
         return self.request.user.reservations.filter(is_active=True).order_by("-created_at")
-
-
-# accounts
-# TODO: refactor after implementing MVP
-class CustomLoginView(LoginView):
-    template_name = "registration/login.html"
-
-    def get_success_url(self):
-        messages.success(self.request, f"Welcome back, {self.request.user.username}!")
-        return reverse_lazy("reservation:cafe-list")
-
-
-class CustomLogoutView(LogoutView):
-    next_page = reverse_lazy("logout")
-    template_name = "registration/logout.html"
-
-
-class SignUpView(SuccessMessageMixin, CreateView):
-    form_class = VisitorCreationForm
-    # do i actually need to redirect user to login after creation ?
-    success_url = reverse_lazy("login")
-    template_name = "registration/signup.html"
-    success_message = "Congratulations, your account was created. Please login to continue."
-
-
-class VisitorUpdateView(LoginRequiredMixin, SuccessMessageMixin, UpdateView):
-    model = Visitor
-    form_class = VisitorUpdateForm
-    template_name = "registration/profile.html"
-    success_url = reverse_lazy("reservation:profile")
-    success_message = "Your profile was updated successfully!"
-
-    def get_object(self):
-        return self.request.user
