+++ templates/base.html
@@ -47,7 +47,7 @@
                         </a>
                         <ul class="dropdown-menu dropdown-menu-end">
                             <li>
-                                <form method="post" action="{% url 'logout' %}" class="d-inline">
+                                <form method="post" action="{% url 'accounts:logout' %}" class="d-inline">
                                     {% csrf_token %}
                                     <button type="submit" class="dropdown-item">
                                         <i class="bi bi-box-arrow-right me-2"></i> Logout
@@ -59,12 +59,12 @@
                 {% else %}
                     <!-- Login/Signup buttons -->
                     <li class="nav-item">
-                        <a class="nav-link" href="{% url 'login' %}">
+                        <a class="nav-link" href="{% url 'accounts:login' %}">
                             <i class="bi bi-box-arrow-in-right me-1"></i> Login
                         </a>
                     </li>
                     <li class="nav-item">
-                        <a class="btn btn-primary ms-2" href="{% url 'signup' %}">
+                        <a class="btn btn-primary ms-2" href="{% url 'accounts:signup' %}">
                             <i class="bi bi-person-plus me-1"></i> Sign Up
                         </a>
                     </li>
