+++ templates/registration/login.html
@@ -18,7 +18,7 @@
                 <hr>
             </div>
             <div class="card-footer text-center">
-                Don't have an account? <a href="{% url "signup" %}">Sign up</a>
+                Don't have an account? <a href="{% url "accounts:signup" %}">Sign up</a>
             </div>
         </div>
     </div>
