+++ templates/registration/logout.html
@@ -9,7 +9,7 @@
             </div>
             <div class="card-body text-center">
                 <p>You have been successfully logged out.</p>
-                <a href="{% url "login" %}" class="btn btn-primary">Login Again</a>
+                <a href="{% url "accounts:login" %}" class="btn btn-primary">Login Again</a>
             </div>
         </div>
     </div>
