+++ templates/registration/profile.html
@@ -39,7 +39,7 @@
                         <button type="submit" class="btn btn-primary">
                             <i class="bi bi-check-circle"></i> Update Profile
                         </button>
-                        <a href="{% url 'profile' %}" class="btn btn-outline-secondary ms-2">
+                        <a href="{% url 'accounts:profile' %}" class="btn btn-outline-secondary ms-2">
                             <i class="bi bi-x-circle"></i> Cancel
                         </a>
                     </div>
