+++ templates/registration/signup.html
@@ -17,7 +17,7 @@
                 </form>
                 <hr>
                 <div class="text-center">
-                    Already have an account? <a href="{% url "login" %}">Login</a>
+                    Already have an account? <a href="{% url "accounts:login" %}">Login</a>
                 </div>
             </div>
         </div>
